backend/pyproject.toml
[tool.poetry]
name = "sistema_portuario"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6"
django = "^2.2.6"
djangorestframework = "^3.10.3"
django-environ = "^0.4.5"
drf-yasg = "^1.17.0"
pycpfcnpj = "^1.5.1"
djangorestframework_simplejwt = "^4.3.0"
psycopg2-binary = "^2.8.4"

[tool.poetry.dev-dependencies]
black = "19.3b0"
flake8 = "^3.7.9"
flake8-bugbear = "^19.8.0"
isort = "^4.3.21"
mimesis = "^3.3.0"
yaspin = "^0.15.0"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
